Move headers to internal directory
