+++ clients/CMakeLists.txt
@@ -40,6 +40,7 @@ if( BUILD_CLIENTS_TESTS OR BUILD_CLIENTS_BENCHMARKS OR BUILD_CLIENTS_SAMPLES )
     add_library(hipsolver_fortran_client ${hipsolver_f90_source_clients})
     add_dependencies(hipsolver_fortran_client hipsolver_fortran)
     include_directories(${CMAKE_BINARY_DIR}/include)
+    include_directories(${CMAKE_BINARY_DIR}/include/internal)
   endif( )
 endif( )
