+++ library/include/internal/hipsolver-compat.h
