+++ library/include/internal/hipsolver-functions.h
