+++ library/include/internal/hipsolver-types.h
